n+1 problem fix

+++ app/Http/Controllers/CategoryController.php
@@ -2,16 +2,21 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Category;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
+use App\Models\Category;
 
+use Illuminate\Support\Facades\Auth;
 
 class CategoryController extends Controller
 {
     public function index()
     {
-        $categories = Category::with('todos')->get();
+        $categories = Category::withCount(['todos' => function ($query) {
+            $query->where('user_id', Auth::id()); 
+        }])
+        ->where('user_id', Auth::id()) 
+        ->get();
+
         return view('category.index', compact('categories'));
     }
 
@@ -28,11 +33,10 @@ class CategoryController extends Controller
 
         Category::create([
             'user_id' => Auth::id(),
-,
             'title' => $request->title,
         ]);
 
-        return redirect()->route('category.index')->with('success', 'Category created successfully.');
+        return redirect()->route('category.index')->with('success', 'Category created successfully!');
     }
 
     public function edit(Category $category)
@@ -50,12 +54,16 @@ class CategoryController extends Controller
             'title' => $request->title,
         ]);
 
-        return redirect()->route('category.index')->with('success', 'Category updated successfully.');
+        return redirect()->route('category.index')->with('success', 'Category updated successfully!');
     }
 
     public function destroy(Category $category)
     {
-        $category->delete();
-        return back()->with('success', 'Category deleted successfully.');
+        if (Auth::id() == $category->user_id) {
+            $category->delete();
+            return redirect()->route('category.index')->with('success', 'Category deleted successfully!');
+        } else {
+        return redirect()->route('category.index')->with('danger', 'You are not authorized to delete this category!');
+        }
     }
 }
